Move TreasuryBenchmarkHelper to common and add to mainnet
